+++ src/main/java/br/com/livelo/orderflight/constants/SkuConstant.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -15,8 +16,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 @ConfigurationProperties(prefix = "fixed-values.sku")
 public class SkuConstant implements Serializable {
-    private Double listPrice;
-    private Double salePrice;
+    private BigDecimal listPrice;
+    private BigDecimal salePrice;
     private Integer quantity;
     private String currency;
     private String description;
