+++ src/main/java/br/com/livelo/orderflight/domain/dtos/installment/InstallmentDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,7 +18,7 @@ public class InstallmentDTO {
 
     private String currency;
 
-    private double amount;
+    private BigDecimal amount;
 
-    private double interest;
+    private BigDecimal interest;
 }
